length modifiers

+++ main.h
@@ -24,7 +24,7 @@ int _prints_hex(unsigned int numb, char value);
 int _prints_nospace(char *str);
 int _nospace(int numb);
 int print_length(long num);
-int print_height(short num);
+int print_height(int num);
 int _prints_plus(int numb, char value);
 int _prints_space(int numb, char value);
 int _prints_hash(unsigned int numb, char value);
